# apps/form-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ConditionDTO:
    properties:
      operator:
        enum:
        - equals
        - not_equals
        - contains
        - not_contains
        - greater_than
        - less_than
        example: equals
        type: string
      questionId:
        example: q1
        type: string
      value:
        example: "yes"
        type: string
    required:
    - operator
    - questionId
    type: object
  dto.ConditionalLogicDTO:
    properties:
      hideIf:
        items:
          $ref: '#/definitions/dto.ConditionDTO'
        type: array
      logic:
        enum:
        - AND
        - OR
        example: AND
        type: string
      showIf:
        items:
          $ref: '#/definitions/dto.ConditionDTO'
        type: array
    type: object
  dto.CreateFormRequestDTO:
    properties:
      allowMultiple:
        example: false
        type: boolean
      category:
        example: feedback
        maxLength: 100
        type: string
      description:
        example: Please provide your feedback about our service
        maxLength: 1000
        type: string
      expiresAt:
        example: "2024-12-31T23:59:59Z"
        type: string
      isAnonymous:
        example: false
        type: boolean
      isPublic:
        example: true
        type: boolean
      questions:
        items:
          $ref: '#/definitions/dto.CreateQuestionRequestDTO'
        minItems: 1
        type: array
      settings:
        $ref: '#/definitions/dto.FormSettingsDTO'
      tags:
        items:
          type: string
        maxItems: 10
        type: array
      title:
        example: Customer Feedback Form
        maxLength: 255
        minLength: 1
        type: string
    required:
    - questions
    - title
    type: object
  dto.CreateQuestionRequestDTO:
    properties:
      conditional:
        $ref: '#/definitions/dto.ConditionalLogicDTO'
      description:
        example: Please enter your full name
        maxLength: 1000
        type: string
      label:
        example: What is your name?
        maxLength: 500
        minLength: 1
        type: string
      metadata:
        additionalProperties: true
        type: object
      options:
        items:
          $ref: '#/definitions/dto.QuestionOptionDTO'
        type: array
      order:
        example: 1
        minimum: 0
        type: integer
      required:
        example: true
        type: boolean
      type:
        enum:
        - text
        - textarea
        - number
        - email
        - date
        - checkbox
        - radio
        - select
        - file
        example: text
        type: string
      validation:
        $ref: '#/definitions/dto.QuestionValidationDTO'
    required:
    - label
    - type
    type: object
...
  dto.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      metadata:
        additionalProperties: true
        type: object
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      correlationId:
        type: string
      error:
        $ref: '#/definitions/dto.ErrorDetail'
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
      version:
        type: string
    type: object
  dto.FormResponseDTO:
    properties:
      allowMultiple:
        example: false
        type: boolean
      category:
        example: feedback
        type: string
      createdAt:
        example: "2024-01-01T12:00:00Z"
        type: string
      createdBy:
        $ref: '#/definitions/dto.UserInfoDTO'
      description:
        example: Please provide your feedback
        type: string
      expiresAt:
        example: "2024-12-31T23:59:59Z"
        type: string
      id:
        example: f123e4567-e89b-12d3-a456-426614174000
        type: string
      isAnonymous:
        example: false
        type: boolean
      isPublic:
        example: true
        type: boolean
      publishedAt:
        example: "2024-01-01T12:00:00Z"
        type: string
      questions:
        items:
          $ref: '#/definitions/dto.QuestionResponseDTO'
        type: array
      settings:
        $ref: '#/definitions/dto.FormSettingsDTO'
      statistics:
        $ref: '#/definitions/dto.FormStatisticsDTO'
      status:
        example: published
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        example: Customer Feedback Form
        type: string
      updatedAt:
        example: "2024-01-01T12:00:00Z"
        type: string
    type: object
  dto.FormSettingsDTO:
    properties:
      allowDrafts:
        example: false
        type: boolean
      collectEmail:
        example: true
        type: boolean
      customCss:
        maxLength: 5000
        type: string
      metadata:
        additionalProperties: true
        type: object
      notifyOnSubmission:
        example: true
        type: boolean
      redirectUrl:
        maxLength: 500
        type: string
      requireLogin:
        example: false
        type: boolean
      showProgressBar:
        example: true
        type: boolean
      thankYouMessage:
        maxLength: 1000
        type: string
    type: object
  dto.FormStatisticsDTO:
    properties:
      averageTimeSeconds:
        example: 120
        type: integer
      completionRate:
        example: 0.85
        type: number
      lastResponse:
        example: "2024-01-01T12:00:00Z"
        type: string
      responseRate:
        example: 0.75
        type: number
      totalResponses:
        example: 150
        type: integer
      uniqueResponders:
        example: 140
        type: integer
    type: object
  dto.FormSummaryDTO:
    properties:
      createdAt:
        example: "2024-01-01T12:00:00Z"
        type: string
      createdBy:
        $ref: '#/definitions/dto.UserInfoDTO'
      description:
        example: Please provide your feedback
        type: string
      expiresAt:
        example: "2024-12-31T23:59:59Z"
        type: string
      id:
        example: f123e4567-e89b-12d3-a456-426614174000
        type: string
      isPublic:
        example: true
        type: boolean
      publishedAt:
        example: "2024-01-01T12:00:00Z"
        type: string
      statistics:
        $ref: '#/definitions/dto.FormStatisticsDTO'
      status:
        example: published
        type: string
      title:
        example: Customer Feedback Form
        type: string
      updatedAt:
        example: "2024-01-01T12:00:00Z"
        type: string
    type: object
...
  dto.PaginatedResponse:
    properties:
      correlationId:
        type: string
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/dto.Pagination'
      success:
        type: boolean
      timestamp:
        type: string
      version:
        type: string
    type: object
  dto.Pagination:
    properties:
      hasNext:
        type: boolean
      hasPrev:
        type: boolean
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  dto.QuestionOptionDTO:
    properties:
      label:
        example: Option 1
        maxLength: 255
        type: string
      order:
        example: 1
        minimum: 0
        type: integer
      value:
        example: option1
        maxLength: 255
        type: string
    required:
    - label
    - value
    type: object
  dto.QuestionResponseDTO:
    properties:
      conditional:
        $ref: '#/definitions/dto.ConditionalLogicDTO'
      createdAt:
        example: "2024-01-01T12:00:00Z"
        type: string
      description:
        example: Please enter your full name
        type: string
      id:
        example: q123e4567-e89b-12d3-a456-426614174000
        type: string
      label:
        example: What is your name?
        type: string
      metadata:
        additionalProperties: true
        type: object
      options:
        items:
          $ref: '#/definitions/dto.QuestionOptionDTO'
        type: array
      order:
        example: 1
        type: integer
      required:
        example: true
        type: boolean
      type:
        example: text
        type: string
      updatedAt:
        example: "2024-01-01T12:00:00Z"
        type: string
      validation:
        $ref: '#/definitions/dto.QuestionValidationDTO'
    type: object
  dto.QuestionValidationDTO:
    properties:
      allowedTypes:
        example:
        - jpg
        - png
        - pdf
        items:
          type: string
        type: array
      maxFileSize:
        example: 5242880
        type: integer
      maxLength:
        example: 100
        type: integer
      maxValue:
        example: 100
        type: number
      minLength:
        example: 2
        type: integer
      minValue:
        example: 0
        type: number
      pattern:
        example: ^[a-zA-Z ]+$
        type: string
    type: object
  dto.SuccessResponse:
    properties:
      correlationId:
        type: string
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
      version:
        type: string
    type: object
...
  dto.UserInfoDTO:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: u123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: John Doe
        type: string
      username:
        example: john_doe
        type: string
    type: object
  dto.ValidationErrorDTO:
    properties:
      correlationId:
        type: string
      error:
        $ref: '#/definitions/dto.ValidationErrorDetail'
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
      version:
        type: string
    type: object
  dto.ValidationErrorDetail:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      fields:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: Input validation failed
        type: string
      metadata:
        additionalProperties: true
        type: object
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: form-service@example.com
    name: Form Service Team
  description: Comprehensive form management service built with Clean Architecture
    and microservices best practices
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Form Service API
  version: 1.0.0
paths:
  /forms:
    get:
      consumes:
      - application/json
      description: Get a paginated list of forms with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by status
        enum:
        - draft
        - published
        - closed
        - archived
        in: query
        name: status
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Search in title and description
        in: query
        name: search
        type: string
      - default: created_at
        description: Sort field
        enum:
        - created_at
        - updated_at
        - title
        - response_count
        in: query
        name: sortBy
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      - description: Filter by creator user ID
        in: query
        name: createdBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Forms retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.FormSummaryDTO'
                  type: array
              type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List forms
      tags:
      - Forms
    post:
      consumes:
      - application/json
      description: Create a new form with questions and settings
      parameters:
      - description: Form creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFormRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Form created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.FormResponseDTO'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/dto.ValidationErrorDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new form
      tags:
      - Forms
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"